完善股票信息页面

+++ shares/src/components/Shares.vue
@@ -30,14 +30,25 @@
 </template>
 
 <script>
-
   import SqlHelper from '../libs/SqlHelper';
 
   export default {
     name: "Shares",
     mounted() {
       // 从数据库中加载股票信息
-      this.datas = SqlHelper.getShares();
+      this.datas = [];
+      let promise = SqlHelper.getShares();
+      if (promise) {
+        promise.then(result => {
+          for (let i = 0; i < result.rows.length; i++) {
+            this.datas.push({id: result.rows.item(i).id, name: result.rows.item(i).name});
+          }
+
+          this.datas.push({id: '', name: ''});
+        }, error => {
+          this.showErrorMsg('加载股票信息失败');
+        })
+      }
     },
     data() {
       return {
@@ -69,13 +80,52 @@
         this.editIndex = index;
         this.editMode = mode;
       },
-      handleDelete (row, index) {
-        this.datas.splice(index, 1);
 
-        // 从数据库中删除数据
-        SqlHelper.deleteShares(row.id);
+      handleDelete (row, index) {
+        let self = this;
+        this.$Modal.confirm({
+          title: '删除股票信息',
+          content: '确认要删除该股票信息吗?',
+          onOk: function () {
+            self.datas.splice(index, 1);
+
+            // 从数据库中删除数据
+            let promise = SqlHelper.deleteShares(row.id);
+            if (promise) {
+              promise.then(result => {
+                self.showSuccessMsg('删除数据成功');
+              }, err => {
+                self.showErrorMsg('删除数据失败');
+              })
+            }
+          }
+        });
       },
-      handleSave (index) {
+
+      handleAdd(index) {
+        let promise = SqlHelper.getSharesById(this.editId);
+        promise.then(result => {
+          if (result.rows.length > 0) {
+            this.showErrorMsg('id重复，请重新输入');
+            return;
+          } else {
+            this.handleSave(index);
+          }
+        });
+      },
+
+      handleSave(index) {
+        let promise = SqlHelper.getSharesById(this.editId);
+        promise.then(result => {
+          if (this.editMode === 2 || result.rows.length <= 0) {
+            this.handleSaveImpl(index);
+          } else if (this.editMode === 1 && result.rows.length > 0) {
+            this.showErrorMsg('id重复，请重新输入');
+          }
+        });
+      },
+
+      handleSaveImpl(index) {
         this.datas[index].id = this.editId;
         this.datas[index].name = this.editName;
         this.editIndex = -1;
@@ -85,12 +135,27 @@
         }
 
         // 保存数据至数据库
+        let promise = null;
         if (this.editMode == 1) {
-          let result = SqlHelper.insertShares(this.editId, this.editName);
-          alert(result);
+          promise = SqlHelper.insertShares(this.editId, this.editName);
         } else {
-          SqlHelper.updateShares(this.editId, this.editName);
+          promise = SqlHelper.updateShares(this.editId, this.editName);
         }
+        if (promise) {
+          promise.then(result => {
+            this.showSuccessMsg(this.editMode == 1 ? '新增数据成功' : '修改数据成功');
+          }, err => {
+            this.showErrorMsg(this.editMode == 1 ? '新增数据失败' : '修改数据失败');
+          })
+        }
+      },
+
+      showSuccessMsg(content) {
+        this.$Message.success({content: content, duration: 5, closable: true, background: true});
+      },
+
+      showErrorMsg(content) {
+        this.$Message.error({content: content, duration: 5, closable: true, background: true});
       },
     }
   }
